+++ asyncmy/replication/row_events.py
@@ -40,7 +40,6 @@ from .constants import (
 from .errors import TableMetadataUnavailableError
 from .events import BinLogEvent
 from .table import Table
-from .utils import byte2int
 
 
 class RowsEvent(BinLogEvent):
@@ -540,10 +539,10 @@ class TableMapEvent(BinLogEvent):
         self.flags = struct.unpack("<H", self.packet.read(2))[0]
 
         # Payload
-        self.schema_length = byte2int(self.packet.read(1))
+        self.schema_length = self.packet.read(1)
         self.schema = self.packet.read(self.schema_length).decode()
         self.packet.advance(1)
-        self.table_length = byte2int(self.packet.read(1))
+        self.table_length = self.packet.read(1)
         self.table_name = self.packet.read(self.table_length).decode()
         schema_table = f"{self.schema}.{self.table_name}"
         if self._only_tables is not None and schema_table not in self._only_tables:
@@ -606,7 +605,7 @@ class TableMapEvent(BinLogEvent):
                         "COLUMN_TYPE": "BLOB",  # we don't know what it is, so let's not do anything with it.
                         "COLUMN_KEY": "",
                     }
-                col = Column(byte2int(column_type), column_schema, self.packet)
+                col = Column(column_type, column_schema, self.packet)
                 self.columns.append(col)
 
         self._table = Table(